Finished Theorem 4.9.5: Full funext from naïve funext.
This shows that f = g and f ~ g are equivalent, i.e. happly is an
equivalence. However, the inverse obtained from this equivalence,
funext', is not the same as the funext we have used, so the homotopies
are currently useless.
Since the only place fun_uniq is used is in funIsSet, this isn't too
much of a worry, I think.